+++ lib/src/observable/observable_list.dart
@@ -53,10 +53,29 @@ class ObservableList<T> extends ObservableValue<List<T>> {
     notifyListeners();
   }
 
+  /// insert element in the index
+  void insert(int index, T element) {
+    this.value?.insert(index, element);
+    notifyListeners();
+  }
+
+  /// insert element list in the index
+  void insertAll(int index, Iterable<T> iterable) {
+    this.value?.insertAll(index, iterable);
+    notifyListeners();
+  }
+
   /// replace elements in list and notify listeners
   void replaceRange(int start, int end, Iterable<T> replacement) {
     this.value?.replaceRange(start, end, replacement);
     notifyListeners();
+  }
+
+  void modifyItem(int index, T modify(T value)) {
+    if (this.value != null && this.value.length > index) {
+      this.value[index] = modify(this.value[index]);
+      notifyListeners();
+    }
   }
 
   /// get if list is empty